render-deploy/requirements-render.txt
# Основные зависимости
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0

# AI/ML APIs (оптимизированные версии)
groq==0.4.1

# Веб-интерфейс (легковесный)
Flask==3.0.0
Flask-CORS==4.0.0

# Обработка данных (минимальные версии)
numpy==1.24.3

# Семантический поиск (опционально)
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Утилиты
colorama==0.4.6
Pillow==10.1.0

# Кэширование
redis==5.0.1